analysis refactoring : moving optimal_extent and insufficient_overlap_message
